+++ src/System/hdf5_readwrite.cc
@@ -89,7 +89,7 @@ namespace Loci {
              H5P_DEFAULT, data);
 
     eset = EMPTY;
-    for(unsigned int i=0;i< dimension;i++){
+    for(size_t i=0;i< dimension;i++){
       eset |= interval(data[i],data[i+1]);
       i++;
     }
