+++ src/System/visit_tools.cc
@@ -110,7 +110,7 @@ namespace Loci {
         for(vmsi=ruleIter->get_info().desc.targets.begin();
             vmsi!=ruleIter->get_info().desc.targets.end(); ++vmsi) {
           if(vmsi->assign.size() != 0)
-            for(unsigned int i=0;i<vmsi->assign.size();++i) {
+            for(size_t i=0;i<vmsi->assign.size();++i) {
               variable new_name = vmsi->assign[i].first ;
               variable old_name = vmsi->assign[i].second ;
